refactor(cli): move profile helpers to utils module
